feat: add routing for interaction-widgets and adjust widget-project-structure
