+++ lib/widgets/generic/adjust_height_slider.dart
